# src/main/java/fr/inria/diverse/todoapp/model/Author.java
package fr.inria.diverse.todoapp.model;

import jakarta.persistence.Id;

public class Author {
    @Id
    private String id;
    private String name;

    public Author() {
    }

    public Author(String id, String name) {
        this.id = id;
        this.name = name;
    }
    public String getId() {
        return id;
    }
    public void setId(String id) {
        this.id = id;
    }
    public String getName() {
        return name;
    }
    public void setName(String name) {
        this.name = name;
    }



    
}