# src/main/java/fr/inria/diverse/todoapp/model/Tag.java
package fr.inria.diverse.todoapp.model;

import jakarta.persistence.Entity;
import jakarta.persistence.Id;

@Entity
public class Tag {
    @Id
    private String id;
    private String name;

    public Tag() {
    }

    public Tag(String id, String name) {
        this.id = id;
        this.name = name;
    }
    public String getId() {
        return id;
    }
    public void setId(String id) {
        this.id = id;
    }
    public String getName() {
        return name;
    }
    public void setName(String name) {
        this.name = name;
    }


    
}